微博展示功能

+++ app/src/main/java/com/snail/iweibo/mvp/model/UserBean.java
@@ -1,10 +1,37 @@
 package com.snail.iweibo.mvp.model;
 /**
  * UserBean
+ * id : 1404376560
+ * screen_name : zaku
+ * name : zaku
+ * province : 11
+ * city : 5
+ * location : 北京 朝阳区
+ * description : 人生五十年，乃如梦如幻；有生斯有死，壮士复何憾。
+ * url : http://blog.sina.com.cn/zaku
+ * profile_image_url : http://tp1.sinaimg.cn/1404376560/50/0/1
+ * domain : zaku
+ * gender : m
+ * followers_count : 1204
+ * friends_count : 447
+ * statuses_count : 2908
+ * favourites_count : 0
+ * created_at : Fri Aug 28 00:00:00 +0800 2009
+ * following : false
+ * allow_all_act_msg : false
+ * remark :
+ * geo_enabled : true
+ * verified : false
+ * allow_all_comment : true
+ * avatar_large : http://tp1.sinaimg.cn/1404376560/180/0/1
+ * verified_reason :
+ * follow_me : false
+ * online_status : 0
+ * bi_followers_count : 215
  * Created by alexwan on 16/3/23.
  */
 public class UserBean {
-    private int id;
+    private long id;
     private String screen_name;
     private String name;
     private String province;
@@ -86,7 +113,7 @@ public class UserBean {
 
     public void setBi_followers_count(int bi_followers_count) { this.bi_followers_count = bi_followers_count;}
 
-    public int getId() { return id;}
+    public long getId() { return id;}
 
     public String getScreen_name() { return screen_name;}
